fix(layout): 헤더 및 사이드 레이아웃 수정
max-width: 144rem 설정
layout 폴더로 파일들 이동
main section을 사이드 nav 옆으로 이동
