--- components/common/Header/innerNavBarModule.ts
@@ -1,40 +1,0 @@
-import styled from "styled-components";
-
-export const inner_menu_bar = styled.nav`
-  display: flex;
-`;
-
-export const vertical_line = styled.nav`
-  border-left: 1px solid #E1E1E1;
-  height: 80vh; /* 원하는 높이 설정 */
-  justify-content: center;
-  align-items: center;
-`;
-
-export const inner_nav_ul = styled.nav`
-  padding  : 5%;
-`;
-
-export const inner_nav_item = styled.li`
-  font-size: 18px;
-  font-weight: 600;
-  color: #979797;
-  margin-top: 90px;
-  list-style: none;
-`;
-export const menuIcon = styled.div`
-  width: 100vw;
-`
-export const LogoBox = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  left: 15%;
-  position: absolute;
-`;
-
-export const ImageWrapper = styled.div`
-  position: absolute;
-`;
-
