+++ layouts/AppLayoutStyles.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+const SideNavBox = styled.div`
+  max-width: 144rem;
+  padding: 2rem;
+  margin: 0 auto;
+  display: flex;
+  flex-direction: row;
+`;
+
+const MainBox = styled.main`
+  padding: 4rem 7rem;
+`
+
+export { SideNavBox, MainBox }
