+++ layouts/InnerNavBar/InnerNavBarModule.ts
@@ -1,0 +1,31 @@
+import styled from "styled-components";
+
+export const inner_menu_bar = styled.nav`
+  width: 18.4rem;
+  height: 75vh;
+  display: flex;
+  border-right: 1px solid #E1E1E1;
+  position: relative;
+`;
+
+export const inner_nav_ul = styled.ul`
+  width: 100%;
+  padding: 2rem;
+  text-align: center;
+`;
+
+export const inner_nav_item = styled.li`
+  font-size: 18px;
+  font-weight: 600;
+  color: #979797;
+  margin-top: 90px;
+  list-style: none;
+`;
+
+export const LogoBox = styled.div`
+  top: 30%;
+  left: 18.4rem;
+  transform: translate(-50%, -50%);
+  position: absolute;
+`;
+
