+++ pages/community/CommunityFriends.tsx
@@ -1,0 +1,8 @@
+/** 커뮤니티 친구 페이지 */
+const CommunityFriends = () => {
+  return (
+    <h1>123</h1>
+  );
+};
+
+export default CommunityFriends;
