+++ pages/login/index.tsx
@@ -1,4 +1,4 @@
-import Header from "../../components/common/Header/header";
+import Header from "../../layouts/Header";
 
 const Login = () => {
   return (
